ttl: Consolidate TTL calculation
Currently part of the TTL calculation is done outside of the
calc_ttl() function.  Ideally a single variable in main()
would have the total seconds until the cache expires, and
ctx.page.expires would be incremented by this, as well
as the parameter to cache_process getting this value in
seconds.

Signed-off-by: Tim Nordell <tim.nordell@logicpd.com>