ttl: Support different TTL times based on cache-control
Allow the client browser to pass in "max-age=0" and "no-cache"
to control a separate TTL time on the server for each type of
page.  This extends the TTL field to have the additional form
of:

    some-ttl=5:1

where 5 is the number of minutes if the user were to simply
load the page, and 1 is the number of minutes if the user were
to reload the page in their web browser (hit refresh).

Signed-off-by: Tim Nordell <tim.nordell@logicpd.com>